# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nexu API",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/internal/sessions": {
      "post": {
        "tags": [
          "Sessions (Internal)"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "botId": {
                    "type": "string",
                    "minLength": 1
                  },
                  "sessionKey": {
                    "type": "string",
                    "minLength": 1
                  },
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 500
                  },
                  "nexuUserId": {
                    "type": "string"
                  },
                  "channelType": {
                    "type": "string"
                  },
                  "channelId": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "ended"
                    ]
                  },
                  "messageCount": {
                    "type": "integer"
                  },
                  "lastMessageAt": {
                    "type": "string"
                  },
                  "metadata": {
                    "type": "object",
                    "additionalProperties": {
                      "nullable": true
                    }
                  }
                },
                "required": [
                  "botId",
                  "sessionKey",
                  "title"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Session created or updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "botId": {
                      "type": "string"
                    },
                    "nexuUserId": {
                      "type": "string",
                      "nullable": true
                    },
                    "sessionKey": {
                      "type": "string"
                    },
                    "channelType": {
                      "type": "string",
                      "nullable": true
                    },
                    "channelId": {
                      "type": "string",
                      "nullable": true
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    },
                    "messageCount": {
                      "type": "number"
                    },
                    "lastMessageAt": {
                      "type": "string",
                      "nullable": true
                    },
                    "metadata": {
                      "type": "object",
                      "nullable": true,
                      "additionalProperties": {
                        "nullable": true
                      }
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "botId",
                    "nexuUserId",
                    "sessionKey",
                    "channelType",
                    "channelId",
                    "title",
                    "status",
                    "messageCount",
                    "lastMessageAt",
                    "metadata",
                    "createdAt",
                    "updatedAt"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/internal/sessions/{id}": {
      "patch": {
        "tags": [
          "Sessions (Internal)"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 500
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "ended"
                    ]
                  },
                  "messageCount": {
                    "type": "integer"
                  },
                  "lastMessageAt": {
                    "type": "string"
                  },
                  "metadata": {
                    "type": "object",
                    "additionalProperties": {
                      "nullable": true
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "botId": {
                      "type": "string"
                    },
                    "nexuUserId": {
                      "type": "string",
                      "nullable": true
                    },
                    "sessionKey": {
                      "type": "string"
                    },
                    "channelType": {
                      "type": "string",
                      "nullable": true
                    },
                    "channelId": {
                      "type": "string",
                      "nullable": true
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    },
                    "messageCount": {
                      "type": "number"
                    },
                    "lastMessageAt": {
                      "type": "string",
                      "nullable": true
                    },
                    "metadata": {
                      "type": "object",
                      "nullable": true,
                      "additionalProperties": {
                        "nullable": true
                      }
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "botId",
                    "nexuUserId",
                    "sessionKey",
                    "channelType",
                    "channelId",
                    "title",
                    "status",
                    "messageCount",
                    "lastMessageAt",
                    "metadata",
                    "createdAt",
                    "updatedAt"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Session not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}